update r2d2 loss function to take into account framestates. +modified a few helping function to deal with hierarchical dictionnaries. --> VERY IMPORTANT: when manipulating hdict, it is important to control when references are being made instead of expected copies. In order to remain sane, a copy_hdict helping function has been implemented. It should now be understood a good practice to use it everytime a hdict is supposed to change scope, e.g. from agents to model inferences. --> TODO: propagate good practices from DQN agent to others. +benchmarking/debugging of r2d2 in progress...

+++ regym/util/wrappers.py
@@ -1363,8 +1363,8 @@ class PreviousRewardActionInfoWrapper(gym.Wrapper):
         info['previous_reward'] = copy.deepcopy(self.previous_reward)
         info['previous_action'] = copy.deepcopy(self.previous_action)
 
-        self.previous_reward = np.ones((1, 1))*reward
-        self.previous_action = np.eye(self.nbr_actions)[action].reshape(1, -1)
+        self.previous_reward = np.ones((1, 1), dtype=np.float32)*reward
+        self.previous_action = np.eye(self.nbr_actions, dtype=np.float32)[action].reshape(1, -1)
 
         return observation, reward, done, info
